+++ src/WeihanLi.EntityFramework/EFRepositoryExtensions.cs
@@ -12,7 +12,7 @@ namespace WeihanLi.EntityFramework
             TEntity entity, CancellationToken cancellationToken)
             where TDbContext : DbContext
             where TEntity : class
-            => repository.UpdateAsync(entity, null, cancellationToken);
+            => repository.UpdateAsync(entity, Array.Empty<string>(), cancellationToken);
 
         public static Task<int> UpdateAsync<TDbContext, TEntity>(this IEFRepository<TDbContext, TEntity> repository,
             TEntity entity,
@@ -28,15 +28,20 @@ namespace WeihanLi.EntityFramework
             where TEntity : class
             => repository.UpdateWithoutAsync(entity, propertyExpressions);
 
-        /// <summary>
-        /// Finds an entity with the given primary key values. If found, is attached to the context and returned. If no entity is found, then null is returned.
-        /// </summary>
-        /// <param name="repository">repository</param>
-        /// <param name="keyValues">The values of the primary key for the entity to be found.</param>
-        /// <returns>A <see cref="Task{TEntity}"/> that represents the asynchronous find operation. The task result contains the found entity or null.</returns>
         public static Task<TEntity> FindAsync<TDbContext, TEntity>(this IEFRepository<TDbContext, TEntity> repository,
             params object[] keyValues)
             where TDbContext : DbContext
-            where TEntity : class => repository.FindAsync(keyValues);
+            where TEntity : class
+        {
+            return repository.FindAsync(keyValues, CancellationToken.None);
+        }
+
+        public static Task<int> DeleteAsync<TDbContext, TEntity>(this IEFRepository<TDbContext, TEntity> repository,
+            params object[] keyValues)
+            where TDbContext : DbContext
+            where TEntity : class
+        {
+            return repository.DeleteAsync(keyValues, CancellationToken.None);
+        }
     }
 }