+++ test/WeihanLi.EntityFramework.Test/EFTestBase.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Xunit;
+
+namespace WeihanLi.EntityFramework.Test
+{
+    public class EFTestBase : IClassFixture<EFTestFixture>
+    {
+        public IServiceProvider Services { get; }
+        public IEFRepository<TestDbContext, TestEntity> Repository { get; }
+
+        public EFTestBase(EFTestFixture fixture)
+        {
+            Services = fixture.Services;
+            Repository = fixture.Services
+                .GetRequiredService<IEFRepository<TestDbContext, TestEntity>>();
+            if (Repository.DbContext.Database.IsInMemory())
+            {
+                Repository.DbContext.Database.EnsureDeleted();
+            }
+            else
+            {
+                Repository.DbContext.Database.EnsureCreated();
+                Repository.DbContext.Database.ExecuteSqlRaw("TRUNCATE TABLE TestEntities");
+            }
+        }
+    }
+}